requirements.txt
about-time==4.2.1 ; python_version >= "3.9" and python_version < "3.12"
absl-py==1.4.0 ; python_version >= "3.9" and python_version < "3.12"
alive-progress==3.1.4 ; python_version >= "3.9" and python_version < "3.12"
autograd==1.6.2 ; python_version >= "3.9" and python_version < "3.12"
cflib==0.1.17 ; python_version >= "3.9" and python_version < "3.12"
chex==0.1.82 ; python_version >= "3.9" and python_version < "3.12"
cloudpickle==2.2.1 ; python_version >= "3.9" and python_version < "3.12"
cma==3.2.2 ; python_version >= "3.9" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12" and sys_platform == "win32" or python_version >= "3.9" and python_version < "3.12" and platform_system == "Windows"
contourpy==1.1.0 ; python_version >= "3.9" and python_version < "3.12"
cycler==0.11.0 ; python_version >= "3.9" and python_version < "3.12"
dataclasses==0.6 ; python_version >= "3.9" and python_version < "3.12"
decorator==5.1.1 ; python_version >= "3.9" and python_version < "3.12"
deprecated==1.2.14 ; python_version >= "3.9" and python_version < "3.12"
dill==0.3.7 ; python_version >= "3.9" and python_version < "3.12"
distrax @ git+https://github.com/deepmind/distrax.git@93c54a8b15c07e57a58b95b519e5510ab99220c9 ; python_version >= "3.9" and python_version < "3.12"
dm-tree==0.1.8 ; python_version >= "3.9" and python_version < "3.12"
etils[epath,epy]==1.4.1 ; python_version >= "3.9" and python_version < "3.12"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "3.11"
expt==0.4.2 ; python_version >= "3.9" and python_version < "3.12"
farama-notifications==0.0.4 ; python_version >= "3.9" and python_version < "3.12"
flax==0.6.11 ; python_version >= "3.9" and python_version < "3.12"
fonttools==4.42.1 ; python_version >= "3.9" and python_version < "3.12"
future==0.18.3 ; python_version >= "3.9" and python_version < "3.12"
gast==0.5.4 ; python_version >= "3.9" and python_version < "3.12"
grapheme==0.6.0 ; python_version >= "3.9" and python_version < "3.12"
gymnasium==0.28.1 ; python_version >= "3.9" and python_version < "3.12"
importlib-metadata==6.8.0 ; python_version >= "3.9" and python_version < "3.10"
importlib-resources==6.0.1 ; python_version >= "3.9" and python_version < "3.12"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "3.12"
jax-dataclasses==1.5.1 ; python_version >= "3.9" and python_version < "3.12"
jax-jumpy==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
jax==0.4.14 ; python_version >= "3.9" and python_version < "3.12"
jaxlib==0.4.14 ; python_version >= "3.9" and python_version < "3.12"
kiwisolver==1.4.5 ; python_version >= "3.9" and python_version < "3.12"
libusb-package==1.0.26.2 ; python_version >= "3.9" and python_version < "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "3.12"
matplotlib==3.7.2 ; python_version >= "3.9" and python_version < "3.12"
mdurl==0.1.2 ; python_version >= "3.9" and python_version < "3.12"
ml-dtypes==0.2.0 ; python_version >= "3.9" and python_version < "3.12"
mplcursors==0.5.2 ; python_version >= "3.9" and python_version < "3.12"
msgpack==1.0.5 ; python_version >= "3.9" and python_version < "3.12"
multiprocess==0.70.15 ; python_version >= "3.9" and python_version < "3.12"
nest-asyncio==1.5.7 ; python_version >= "3.9" and python_version < "3.12"
numpy==1.25.2 ; python_version < "3.12" and python_version >= "3.9"
nvidia-cublas-cu11==11.10.3.66 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Linux"
nvidia-cuda-nvrtc-cu11==11.7.99 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Linux"
nvidia-cuda-runtime-cu11==11.7.99 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Linux"
nvidia-cudnn-cu11==8.5.0.96 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Linux"
opencv-python-headless==4.5.5.64 ; python_version >= "3.9" and python_version < "3.12"
opt-einsum==3.3.0 ; python_version >= "3.9" and python_version < "3.12"
optax==0.1.7 ; python_version >= "3.9" and python_version < "3.12"
orbax-checkpoint==0.3.5 ; python_version >= "3.9" and python_version < "3.12"
packaging==23.1 ; python_version >= "3.9" and python_version < "3.12"
pandas==2.0.3 ; python_version >= "3.9" and python_version < "3.12"
pettingzoo==1.24.0 ; python_version >= "3.9" and python_version < "3.12"
pillow==10.0.0 ; python_version >= "3.9" and python_version < "3.12"
pluggy==1.2.0 ; python_version >= "3.9" and python_version < "3.12"
protobuf==4.24.1 ; python_version >= "3.9" and python_version < "3.12"
pygame==2.5.1 ; python_version >= "3.9" and python_version < "3.12"
pygments==2.16.1 ; python_version >= "3.9" and python_version < "3.12"
pymoo==0.6.0.1 ; python_version >= "3.9" and python_version < "3.12"
pyopengl==3.1.6 ; python_version >= "3.9" and python_version < "3.12"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "3.12"
pytest==7.4.0 ; python_version >= "3.9" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.12"
pytz==2023.3 ; python_version >= "3.9" and python_version < "3.12"
pyusb==1.2.1 ; python_version >= "3.9" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "3.12"
rich==13.5.2 ; python_version >= "3.9" and python_version < "3.12"
scipy==1.11.2 ; python_version >= "3.9" and python_version < "3.12"
setuptools-scm==7.1.0 ; python_version >= "3.9" and python_version < "3.12"
setuptools==68.1.2 ; python_version >= "3.9" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
supersuit==3.9.0 ; python_version >= "3.9" and python_version < "3.12"
tensorflow-probability==0.21.0 ; python_version >= "3.9" and python_version < "3.12"
tensorstore==0.1.41 ; python_version >= "3.9" and python_version < "3.12"
tinyscaler==1.2.6 ; python_version >= "3.9" and python_version < "3.12"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
toolz==0.12.0 ; python_version >= "3.9" and python_version < "3.12"
torch==1.13.1 ; python_version >= "3.9" and python_version < "3.12"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "3.12"
typeguard==4.0.0 ; python_version >= "3.9" and python_version < "3.12"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "3.12"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.12"
wheel==0.41.2 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Linux"
wrapt==1.15.0 ; python_version >= "3.9" and python_version < "3.12"
zipp==3.16.2 ; python_version >= "3.9" and python_version < "3.12"
fsspec==2025.3.2 ; python_version >= "3.9" and python_version < "3.12"
joblib==1.5.1 ; python_version >= "3.9" and python_version < "3.12"
pybullet==3.2.5 ; python_version >= "3.9" and python_version < "3.12"
scikit-learn==1.6.1 ; python_version >= "3.9" and python_version < "3.12"
threadpoolctl==3.6.0 ; python_version >= "3.9" and python_version < "3.12"